// other/src/shared/types/interfaces.ts
export interface Pagination {
  totalRecords: number;
  currentPage: number;
  size: number;
  totalPages: number;
  [key: string]: any; // For any additional pagination metadata
}

export interface SendResponseParams {
  res: Response;
  data?: any;
  message?: string;
  statusCode?: number;
}

export interface SendErrorParams {
  res: Response;
  message?: string;
  statusCode?: number;
  errors?: IError[];
}

export type SummaryKey =
  | "openingQty"
  | "openingAmount"
  | "increaseQty"
  | "increaseAmount"
  | "decreaseQty"
  | "decreaseAmount"
  | "closingQty"
  | "closingAmount"
  | "outOfStockItems"
  | "lowStockItems"
  | "overstockItems"
  // TODO: totalKey
  | "totalRevenue"
  | "totalExpense"
  | "totalIncome"
  | "totalCost"
  | "totalProfit"
  | "totalDebt"
  | "totalReceivableDebt"
  | "totalPayableDebt"
  | "totalIncreaseAmount"
  | "totalDecreaseAmount"
  | "totalInQty"
  | "totalInAmount"
  | "totalOutQty"
  | "totalOutAmount";

export type SummaryData = {
  [key in SummaryKey]?: number;
};
export interface ApiResponse<T = any> {
  statusCode: number;
  success: boolean;
  message: string;
  data: T;
  pagination?: Pagination;
  errors?: any;
  code?: string;
  summary?: SummaryData; // Optional error code for more specific error handling
}